requirements.txt
SpeechRecognition
pyttsx3
sounddevice
soundfile
numpy
pvporcupine
pytest
requests
beautifulsoup4
pillow
pyautogui
keyboard
pyperclip
pystray